refactor: update embedding model name and commit ID
