+++ source/infrastructure/cli/magic-config.ts
@@ -6,13 +6,14 @@ import * as fs from "fs";
 import * as AWS from "aws-sdk";
 import {
   SystemConfig
-} from "./types";
+} from "../lib/shared/types";
 import { LIB_VERSION } from "./version.js";
 
 const embeddingModels = [
   {
     provider: "sagemaker",
-    name: "maidalun1020/bce-embedding-base_v1",
+    name: "bce-embedding-and-bge-reranker",
+    commitId: "43972580a35ceacacd31b95b9f430f695d07dde9",
     dimensions: 1024,
     default: true,
   }
